+++ js/main.js
@@ -1,10 +1,13 @@
 import { drawMiniatures } from './draw-miniatures.js';
 import { getData } from './api.js';
+import { showGetDataErrorMessage } from './util.js';
 import { getPhotos } from './draw-full-images.js';
 import './upload-image-form.js';
 
-getData().then((photos) => {
-  drawMiniatures(photos);
-  getPhotos(photos);
-});
+getData()
+  .then((photos) => {
+    drawMiniatures(photos);
+    getPhotos(photos);
+  })
+  .catch(() => showGetDataErrorMessage());
